requirements.txt
pandas
pyyaml
scikit-learn
sqlalchemy
boto3
